requirements.txt
streamlit
langchain
python-dotenv
chromadb
langchain_openai
langchain_community
langchain_core
sentence_transformers
pypdf
openai
--extra-index-url https://download.pytorch.org/whl/cpu
torch
transformers